# Prioritize task execution by recipe order