--- a/esmvalcore/_task.py
+++ b/esmvalcore/_task.py
@@ -228,8 +228,9 @@ class BaseTask:
                         os.getpid())
             start = datetime.datetime.now()
             self.output_files = self._run(input_files)
-            logger.info("Successfully completed task %s in %s",
-                        self.name, datetime.datetime.now() - start)
+            runtime = datetime.datetime.now() - start
+            logger.info("Successfully completed task %s (priority %s) in %s",
+                        self.name, self.priority, runtime)
 
         return self.output_files