Closes #3