@mattiarighi 2019-06-18
Is this still a draft or can I already try it?

This feature would help me a lot to get more robust number for the CPU time, which I want to report in the paper.