@bouweandela 2019-06-19
You can already try it, i think it works, but I need to change it a bit so the priority of the input variables used for derivation is correctly set and I need to add unit tests.