@mattiarighi 2019-07-10
I finally found some time to test this.

Unfortunately the NCL scripts are crashing due to missing `variable_info`:
```
NCL: fatal:Undefined identifier: (variable_info) is undefined, can't continue
```
That's strange, since the `variable_info` is correctly defined in `<var>_info.ncl` files.

[main_log.txt](https://github.com/ESMValGroup/ESMValCore/files/3376851/main_log.txt)
[NCL log for ta850](https://github.com/ESMValGroup/ESMValCore/files/3376814/log.txt)