@mattiarighi 2019-07-17
Anything I can do to help with this?
It would really be a great feature to run some reliable performance tests.