@mattiarighi 2019-09-18
To fully make use of this feature, it would be useful to write the time required by each task in the debug log (this is done only for the provenance at the moment). In this way, the user can execute the recipe once and then reorder the task according to the runtimes.

Is that possible?