@mattiarighi 2019-09-18
I tried to reorder the diagnostics in `recipe_perfemtrics_CMIP5.yml`, putting the slowest (`sm` and `taZONAL`) at the beginning, and testing with 16 parallel tasks.

The task are now started in the order listed in the recipe, but this helps only for the preprocessor part of the diagnostics. See the list below: the two tasks to derive the variable `sm` for the first diagnostic are started as first, but the other preprocessing of `sm` and the diagnostic script itself `sm/grading` are called later, in a random order. The same happens with `taZONAL`.

```
Starting task sm/sm_derive_input_sm in process [13499]
Starting task sm/sm_derive_input_mrsos in process [13498]
Starting task taZONAL/ta in process [13500]
Starting task ta850/ta in process [13501]
Starting task ta200/ta in process [13502]
Starting task ta30/ta in process [13503]
Starting task ta5/ta in process [13504]
Starting task ua850/ua in process [13505]
Starting task ua200/ua in process [13506]
Starting task va850/va in process [13507]
Starting task va200/va in process [13508]
Starting task zg500/zg in process [13509]
Starting task hus400/hus in process [13510]
Starting task tas/tas in process [13511]
Starting task ts/ts in process [13512]
Starting task pr/pr in process [13513]
Starting task clt/clt in process [13499]
Starting task ta200/cycle in process [13502]
Starting task ta850/cycle in process [13501]
Starting task ta850/grading in process [13504]
Starting task ta200/grading in process [13503]
Starting task ta30/cycle in process [13507]
Starting task ta5/cycle in process [13508]
Starting task sm/sm in process [13498]
Starting task va850/grading in process [13501]
Starting task va200/grading in process [13502]
Starting task ua200/grading in process [13506]
Starting task ua850/grading in process [13505]
Starting task rlut/rlut in process [13508]
Starting task hus400/grading in process [13510]
Starting task rsut/rsut in process [13507]
Starting task lwcre/lwcre_derive_input_rlutcs in process [13503
Starting task lwcre/lwcre_derive_input_rlut in process [13504]
Starting task swcre/swcre_derive_input_rsutcs in process [13510
Starting task zg500/grading in process [13509]
Starting task swcre/swcre_derive_input_rsut in process [13501]
Starting task od550aer/od550aer in process [13502]
Starting task od870aer/od870aer in process [13506]
Starting task abs550aer/abs550aer in process [13505]
Starting task pr/grading in process [13513]
Starting task ts/grading in process [13512]
Starting task clt/latlon in process [13499]
Starting task clt/grading in process [13509]
Starting task sm/grading in process [13498]
Starting task od550lt1aer/od550lt1aer in process [13513]
Starting task tas/latlon in process [13511]
Starting task abs550aer/grading in process [13505]
Starting task tas/grading in process [13512]
Starting task od870aer/grading in process [13506]
Starting task toz/toz_derive_input_ps in process [13506]
Starting task toz/toz_derive_input_toz in process [13505]
Starting task toz/toz_derive_input_tro3 in process [13505]
Starting task od550lt1aer/grading in process [13506]
Starting task toz/toz in process [13513]
Starting task od550aer/grading in process [13505]
Starting task toz/grading_global in process [13502]
Starting task toz/grading_shpolar in process [13509]
Starting task rsut/grading in process [13506]
Starting task rlut/grading in process [13512]
Starting task lwcre/lwcre in process [13505]
Starting task swcre/swcre in process [13511]
Starting task lwcre/latlon in process [13513]
Starting task lwcre/grading in process [13509]
Starting task swcre/latlon in process [13502]
Starting task swcre/grading in process [13507]
Starting task taZONAL/zonal in process [13504]
Starting task collect/RMSD in process [13498]
Starting task collect/taylor in process [13508]
```

So this PR helps a lot and should be merged, but it's probably not addressing #3 completely.
The results of 5 benchmark tests with and without this feature are shown below: there is definitely an improvement, but the variability is still there.

Without task priority:
0:25:33.422961
0:23:45.208164
0:21:42.018953
0:23:50.358184
0:19:26.837539
**mean=22m 51s**
**median=23m 45s**

With task priority:
0:17:19.614452
0:17:47.433762
0:22:01.984059
0:22:51.531649
0:25:54.753893
**mean=21m 11s**
**median=22m 1s**