@bouweandela 2019-09-18
I added the run time and task priority to the info message, this should help with the ordering.

>  the two tasks to derive the variable `sm` for the first diagnostic are started as first, but the other preprocessing of `sm` and the diagnostic script itself `sm/grading` are called later, in a random order. The same happens with `taZONAL`.

This is because the current implementation should start as many tasks as configured in config-user.yml, starting with the tasks with the lowest priority. However, it can only start a task if all of it's ancestors are ready. Could you post the results above, but including the task completed messages? Because then we can see if there is a problem or tasks are simply waiting for ancestor tasks.